vehicles/vehicleApp/serializers.py
from rest_framework import serializers
from .models import Vehicle

class VehicleListSerializer(serializers.ModelSerializer):
    class Meta:
        model = Vehicle
        fields = ['id', 'name', 'passenger']

class VehicleCreateSerializer(serializers.ModelSerializer):
    class Meta:
        model = Vehicle
        fields = ['id', 'name', 'description', 'passenger']